LayoutTests/css3/filters/backdrop/backdropfilter-property-parsing.html

<!DOCTYPE>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Test the parsing of the -webkit-backdrop-filter property.");

function jsWrapperClass(node)
{
    if (!node)
        return "[null]";
    var string = Object.prototype.toString.apply(node);
    return string.substr(8, string.length - 9);
}

function shouldBeType(expression, className)
{
    shouldBe("jsWrapperClass(" + expression + ")", "'" + className + "'");
    shouldBe("jsWrapperClass(" + expression + ".__proto__)", "'" + className + "'");
    shouldBe("jsWrapperClass(" + expression + ".constructor)", "'Function'");
}

// These have to be global for the test helpers to see them.
var stylesheet, cssRule, declaration
var styleElement = document.createElement("style");
document.head.appendChild(styleElement);
stylesheet = styleElement.sheet;

function testFilterRule(description, rule, expectedValue)
{
    debug("");
    debug(description + " : " + rule);

    stylesheet.insertRule("body { -webkit-backdrop-filter: " + rule + "; }", 0);
    cssRule = stylesheet.cssRules.item(0);

    shouldBe("cssRule.type", "1");

    declaration = cssRule.style;
    shouldBe("declaration.length", "1");
    shouldBeEqualToString("declaration.getPropertyValue('-webkit-backdrop-filter')", expectedValue || rule);

    stylesheet.deleteRule(0);
}

function testInvalidFilterRule(description, rule)
{
    debug("");
    debug(`${description} : ${rule}`);

    stylesheet.insertRule("body { -webkit-backdrop-filter: " + rule + "; }", 0);
    cssRule = stylesheet.cssRules.item(0);

    shouldBe("cssRule.type", "1");

    declaration = cssRule.style;
    shouldBe("declaration.length", "0");
    shouldBe("declaration.getPropertyValue('-webkit-backdrop-filter')", "''");

    stylesheet.deleteRule(0);
}

testFilterRule("Basic reference",
               "url('#a')", 'url("#a")')

testFilterRule("Bare unquoted reference converting to quoted form",
               "url(#a)", 'url("#a")')

testFilterRule("Multiple references",
               "url('#a') url('#b')", 'url("#a") url("#b")')

testFilterRule("Reference as 2nd value",
               "grayscale(1) url('#a')", 'grayscale(1) url("#a")')

testFilterRule("Integer value",
               "grayscale(1)")

testFilterRule("Percentage value",
               "grayscale(50%)")

testFilterRule("Float value converts to integer",
               "grayscale(1.0)", "grayscale(1)")

testFilterRule("Values over 1 are clamped",
               "grayscale(1.5)", "grayscale(1)")

testFilterRule("Percentages over 100 are clamped",
               "grayscale(320%)", "grayscale(100%)")

testFilterRule("Zero value",
               "grayscale(0)")

testFilterRule("No values",
               "grayscale()")

testFilterRule("Multiple values",
               "grayscale(0.5) grayscale(0.25)")

testFilterRule("Integer value",
               "sepia(1)")

testFilterRule("Percentage value",
               "sepia(50%)")

testFilterRule("Float value converts to integer",
               "sepia(1.0)", "sepia(1)")

testFilterRule("Values over 1 are clamped",
               "sepia(8)", "sepia(1)")

testFilterRule("Percentages over 100 are clamped",
               "sepia(101%)", "sepia(100%)")

testInvalidFilterRule("Negative value", "sepia(-0.5)");

testFilterRule("Zero value",
               "sepia(0)")

testFilterRule("No values",
               "sepia()")

testInvalidFilterRule("Negative value", "grayscale(-0.2)")

testFilterRule("Multiple values",
               "sepia(0.5) sepia(0.25)")

testFilterRule("Rule combinations",
               "sepia(0.5) grayscale(0.25)")

testFilterRule("Integer value",
               "saturate(1)")

testFilterRule("Percentage value",
               "saturate(50%)")

testFilterRule("Percentage value > 1",
               "saturate(250%)")

testFilterRule("Float value converts to integer",
               "saturate(1.0)", "saturate(1)")

testFilterRule("Input value > 1",
               "saturate(5.5)")

testInvalidFilterRule("Negative value", "saturate(-0.5)")

testFilterRule("Zero value",
               "saturate(0)")

testFilterRule("No values",
               "saturate()")

testFilterRule("Multiple values",
               "saturate(0.5) saturate(0.25)")

testFilterRule("Rule combinations",
               "saturate(0.5) grayscale(0.25)")

testFilterRule("Degrees value as integer",
               "hue-rotate(10deg)")

testFilterRule("Degrees float value converts to integer",
               "hue-rotate(10.0deg)", "hue-rotate(10deg)")

testFilterRule("Radians value",
               "hue-rotate(10rad)")

testFilterRule("Gradians value",
               "hue-rotate(10grad)")

testFilterRule("Turns value",
               "hue-rotate(0.5turn)")

testFilterRule("Negative value",
               "hue-rotate(-370.2deg)")

testFilterRule("Zero value",
               "hue-rotate(0)", "hue-rotate(0deg)")

testFilterRule("No values",
               "hue-rotate()")

testFilterRule("Rule combinations",
               "hue-rotate(10deg) grayscale(0.25)")

testFilterRule("Integer value",
               "invert(1)")

testFilterRule("Percentage value",
               "invert(50%)")

testFilterRule("Float value converts to integer",
               "invert(1.0)", "invert(1)")

testFilterRule("Values over 1 are clamped",
               "invert(1.01)", "invert(1)")

testFilterRule("Percentages over 100 are clamped",
               "invert(500000%)", "invert(100%)")

testFilterRule("Zero value",
               "invert(0)")

testInvalidFilterRule("Negative value", "invert(-0.5)");

testFilterRule("No values",
               "invert()")

testFilterRule("Multiple values",
               "invert(0.5) invert(0.25)")

testFilterRule("Rule combinations",
               "invert(0.5) grayscale(0.25)")

testFilterRule("Integer value",
               "opacity(1)")

testFilterRule("Percentage value",
               "opacity(50%)")

testFilterRule("Float value converts to integer",
               "opacity(1.0)", "opacity(1)")

testFilterRule("Values over 1 are clamped",
               "opacity(2134687326)", "opacity(1)")

testFilterRule("Percentages over 100 are clamped",
               "opacity(500%)", "opacity(100%)")

testInvalidFilterRule("Negative value", "opacity(-0.5)");

testFilterRule("Zero value",
               "opacity(0)")

testFilterRule("No values",
               "opacity()")

testFilterRule("Multiple values",
               "opacity(0.5) opacity(0.25)")

testFilterRule("Rule combinations",
               "opacity(0.5) grayscale(0.25)")

testFilterRule("Integer value",
              "brightness(1)")

testFilterRule("Percentage value",
              "brightness(50%)")

testFilterRule("Float value converts to integer",
              "brightness(1.0)", "brightness(1)")

testFilterRule("Zero value",
              "brightness(0)")

testInvalidFilterRule("Negative value", "brightness(-2)");

testFilterRule("No values",
              "brightness()")

testFilterRule("Multiple values",
              "brightness(0.5) brightness(0.25)")

testFilterRule("Rule combinations",
              "brightness(0.5) grayscale(0.25)")

testInvalidFilterRule("Value less than -100%", "brightness(-1.1)");
testInvalidFilterRule("Negative value", "brightness(-0.6)");

testFilterRule("Parameter more than 100%",
              "brightness(101%)")

testFilterRule("Rule combinations",
              "grayscale(0.25) brightness(0.5)")

testFilterRule("Integer value",
              "contrast(1)")

testFilterRule("Percentage value",
              "contrast(50%)")

testFilterRule("Percentage value > 1",
              "contrast(250%)")

testFilterRule("Float value converts to integer",
              "contrast(1.0)", "contrast(1)")

testFilterRule("Zero value",
              "contrast(0)")

testInvalidFilterRule("Negative value", "contrast(-0.2)");

testFilterRule("No values",
              "contrast()")

testFilterRule("Value greater than one",
              "contrast(2)")

testInvalidFilterRule("Negative value", "contrast(-0.8)");

testFilterRule("Multiple values",
              "contrast(0.5) contrast(0.25)")

testFilterRule("Rule combinations",
              "contrast(0.5) grayscale(0.25)")

testFilterRule("Rule combinations",
              "grayscale(0.25) contrast(0.5)")

testFilterRule("One zero to px",
               "blur(0)", "blur(0px)")

testFilterRule("One length",
               "blur(10px)")

testFilterRule("No values",
               "blur()")

testInvalidFilterRule("Negative value", "blur(-2px)");

testFilterRule("Color then three values",
              "drop-shadow(red 1px 2px 3px)")

testFilterRule("Three values then color",
              "drop-shadow(1px 2px 3px red)", "drop-shadow(red 1px 2px 3px)")

testFilterRule("Color then three values with zero length",
              "drop-shadow(#abc 0 0 0)", "drop-shadow(rgb(170, 187, 204) 0px 0px 0px)")

testFilterRule("Three values with zero length",
              "drop-shadow(0 0 0)", "drop-shadow(0px 0px 0px)")

testFilterRule("Two values no color",
              "drop-shadow(1px 2px)")

testFilterRule("Multiple operations",
               "grayscale(0.5) sepia(0.25) saturate(0.75) hue-rotate(35deg) invert(0.2) opacity(0.9) blur(5px) drop-shadow(green 1px 2px 3px)");

successfullyParsed = true;

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
